@@ -1,21 +1,28 @@
 apply plugin: 'com.android.application'
+apply plugin: 'kotlin-android'
+apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 28
     defaultConfig {
         applicationId "com.clexi.clexi"
         minSdkVersion 21
-        targetSdkVersion 21
+        targetSdkVersion 27
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+        multiDexEnabled true
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+    }
+    /* This is how to use Java 8 Language Lambda feature */
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
     sourceSets {
         main {
@@ -28,32 +35,48 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    /* Kotlin Configuration */
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${rootProject.ext.kotlinVersion}"
+
+    /* JAR Dependencies */
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+
+    /* Testing Tools */
+    implementation 'junit:junit:4.12'
+    androidTestImplementation('androidx.test.espresso:espresso-core:3.1.1', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support.constraint:constraint-layout:1.0.0-alpha9'
-    compile 'com.android.support:design:25.3.1'
-    testCompile 'junit:junit:4.12'
+
+    /* Supports */
+    implementation 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
+    implementation 'com.google.android.material:material:1.1.0-alpha02'
 
     /* SUPPORTS */
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:design:25.3.1'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile 'com.android.support.constraint:constraint-layout:1.0.0-alpha9'
-    androidTestCompile 'com.android.support:support-annotations:25.3.1'
+    androidTestImplementation 'androidx.annotation:annotation:1.0.1'
+    implementation 'androidx.appcompat:appcompat:1.1.0-alpha01'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha2'
+    implementation 'com.google.android.material:material:1.1.0-alpha02'
+    implementation 'androidx.cardview:cardview:1.0.0'
 
     /* Database */
-    compile 'com.github.satyan:sugar:1.4'
+    implementation 'com.github.satyan:sugar:1.4'
 
     /* Binding Views */
-    compile 'com.jakewharton:butterknife:8.5.1'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.5.1'
+    implementation 'com.jakewharton:butterknife:9.0.0-rc2'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:9.0.0-rc2'
 
     /* View */
-    compile 'com.github.rubensousa:bottomsheetbuilder:1.6.0'
-    compile 'com.sothree.slidinguppanel:library:3.3.1'
-    compile 'com.amulyakhare:com.amulyakhare.textdrawable:1.0.1'
-    compile 'br.com.mauker.materialsearchview:materialsearchview:1.2.2'
+    implementation 'com.github.rubensousa:bottomsheetbuilder:1.6.0'
+    implementation 'com.sothree.slidinguppanel:library:3.3.1'
+    implementation 'com.amulyakhare:com.amulyakhare.textdrawable:1.0.1'
+    implementation 'br.com.mauker.materialsearchview:materialsearchview:1.2.2'
+
+    /* Mirror is a simple DSL to make Java reflection easier. */
+    implementation 'net.vidageek:mirror:1.6.1'
+
+    /**
+     * To add an external maven dependency:
+     * compile 'groupId:artifactId:version'
+     */
 }
