@@ -2,11 +2,6 @@ package com.clexi.clexi.activity;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v7.app.ActionBar;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.ContextMenu;
 import android.view.Menu;
@@ -15,16 +10,22 @@ import android.view.View;
 
 import com.clexi.clexi.R;
 import com.clexi.clexi.app.Consts;
+import com.clexi.clexi.bluetoothle.BleManager;
+import com.clexi.clexi.bluetoothle.object.ApduCommand;
 import com.clexi.clexi.dialog.ConfirmDialog;
-import com.clexi.clexi.dialog.FlyingFOB;
 import com.clexi.clexi.model.access.DbManager;
 import com.clexi.clexi.model.object.Account;
+import com.clexi.clexi.test.TestCommands;
 import com.github.rubensousa.bottomsheetbuilder.BottomSheetBuilder;
 import com.github.rubensousa.bottomsheetbuilder.BottomSheetMenuDialog;
-import com.github.rubensousa.bottomsheetbuilder.adapter.BottomSheetItemClickListener;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import java.util.List;
 
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.widget.Toolbar;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
@@ -57,11 +58,12 @@ public class MainActivity extends BaseActivity
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        //toolbar.setLogo(R.mipmap.ic_logo);
         setSupportActionBar(toolbar);
 
         // Set ActionBar
         ActionBar actionBar = getSupportActionBar();
-        actionBar.setDisplayShowTitleEnabled(true);
+        actionBar.setDisplayShowTitleEnabled(false);
         actionBar.setDisplayHomeAsUpEnabled(false);
 
         // Add some dummy data :)
@@ -74,7 +76,7 @@ public class MainActivity extends BaseActivity
         bindViews();
 
         // Test with simulator
-        startService(new Intent(MainActivity.this, FlyingFOB.class));
+        //startService(new Intent(MainActivity.this, FlyingFOB.class));
     }
 
     @Override
@@ -216,7 +218,27 @@ public class MainActivity extends BaseActivity
             public void onClick(View view)
             {
                 // Add new empty account
-                goToAddAccount();
+                //goToAddAccount();
+
+                // TEST
+                // Request Battery Level
+                ApduCommand command = TestCommands.requestBatteryLevel();
+                if (BleManager.getInstance().isConnected())
+                {
+                    BleManager.getInstance().transmitBlePacketData(com.clexi.clexi.bluetoothle.Consts.UUID_CHARACTERISTIC_CLEXI_REQUEST, new byte[]{
+                            (byte) 0xC0,
+                            (byte) 0x00,
+                            (byte) 0x00,
+                            (byte) 0x07,
+                            (byte) 0x00,
+                            (byte) 0x20,
+                            (byte) 0x00,
+                            (byte) 0x00,
+                            (byte) 0x00,
+                            (byte) 0x00,
+                            (byte) 0x00
+                    });
+                }
             }
         });
 
@@ -262,6 +284,16 @@ public class MainActivity extends BaseActivity
         startActivityForResult(intent, Consts.REQUEST_ADD_EDIT_DELETE_ACCOUNT);
     }
 
+    private void goToViewAccount(long accountId)
+    {
+        Intent intent = new Intent(MainActivity.this, AccountDetailsActivity.class);
+        intent.putExtra(Consts.ACCOUNT_ID, accountId);
+        // With this below code, the second activity return result immediatelly!
+        // So, I commented it for now.
+        //intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_MULTIPLE_TASK);
+        startActivityForResult(intent, Consts.REQUEST_ADD_EDIT_DELETE_ACCOUNT);
+    }
+
     private void goToConnectionSettings()
     {
         Intent intent = new Intent(MainActivity.this, ConnectionSettingsActivity.class);
@@ -273,48 +305,44 @@ public class MainActivity extends BaseActivity
         BottomSheetMenuDialog dialog = new BottomSheetBuilder(this, R.style.AppTheme_BottomSheetDialog)
                 .setMode(BottomSheetBuilder.MODE_LIST)
                 .setMenu(R.menu.menu_account)
-                .setItemClickListener(new BottomSheetItemClickListener()
+                .setItemClickListener(item ->
                 {
-                    @Override
-                    public void onBottomSheetItemClick(final MenuItem item)
+                    if (item.getItemId() == R.id.menu_details)
                     {
-                        if (item.getItemId() == R.id.menu_details)
+                        // Go to AccountDetailsActivity
+                        goToViewAccount(account.getId());
+                    }
+                    else if (item.getItemId() == R.id.menu_edit)
+                    {
+                        // Go to AddAccountActivity for edit
+                        goToEditAccount(account.getId());
+                    }
+                    else if (item.getItemId() == R.id.menu_delete)
+                    {
+                        new ConfirmDialog().setListener(new ConfirmDialog.DialogListener()
                         {
-                            // Go to AccountDetailsActivity
-                            // todo later...
-                        }
-                        else if (item.getItemId() == R.id.menu_edit)
-                        {
-                            // Go to AddAccountActivity for edit
-                            goToEditAccount(account.getId());
-                        }
-                        else if (item.getItemId() == R.id.menu_delete)
-                        {
-                            new ConfirmDialog().setListener(new ConfirmDialog.DialogListener()
+                            @Override
+                            public void onPositiveButtonClicked()
                             {
-                                @Override
-                                public void onPositiveButtonClicked()
-                                {
-                                    // OK, delete the Account
-                                    DbManager.deleteAccount(account);
-                                    mAccounts.remove(account);
+                                // OK, delete the Account
+                                DbManager.deleteAccount(account);
+                                mAccounts.remove(account);
 
-                                    // Update list with dataset to show
-                                    // todo later...
+                                // Update list with dataset to show
+                                // todo later...
 
-                                    mAccountsAdapter.updateList(mAccounts);
-                                    mAccountsAdapter.notifyDataSetChanged();
-                                }
+                                mAccountsAdapter.updateList(mAccounts);
+                                mAccountsAdapter.notifyDataSetChanged();
+                            }
 
-                                @Override
-                                public void onNegativeButtonClicked()
-                                {
-                                    // Nothing
-                                }
-                            })
-                                    .setMessage(String.format(getString(R.string.confirm_delete_account), account.getTitle()))
-                                    .show(getFragmentManager(), "");
-                        }
+                            @Override
+                            public void onNegativeButtonClicked()
+                            {
+                                // Nothing
+                            }
+                        })
+                                .setMessage(String.format(getString(R.string.confirm_delete_account), account.getTitle()))
+                                .show(getFragmentManager(), "");
                     }
                 })
                 .createDialog();
